Refactor FWF processing into a modular package structure with backward compatibility
